# apps/api/v2/swagger/documentation.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v2/gcal/oauth/auth-url": {
      "get": {
        "operationId": "GcalController_redirect",
        "summary": "Get Google Calendar Oauth Url",
        "parameters": [
          {
            "name": "Authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GcalAuthUrlOutput"
                }
              }
            }
          }
        },
        "tags": [
          "Google Calendar"
        ]
      }
    },
    "/v2/gcal/oauth/save": {
      "get": {
        "operationId": "GcalController_save",
        "summary": "Redirect URI for GCal OAuth",
        "parameters": [
          {
            "name": "state",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GcalSaveRedirectOutput"
                }
              }
            }
          }
        },
        "tags": [
          "Google Calendar"
        ]
      }
    },
...
    "/v2/provider/{clientId}": {
      "get": {
        "operationId": "CalProviderController_verifyClientId",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderVerifyClientOutput"
                }
              }
            }
          }
        },
        "tags": [
          "OAuth Cal provider"
        ]
      }
    },
    "/v2/provider/{clientId}/access-token": {
      "get": {
        "operationId": "CalProviderController_verifyAccessToken",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderVerifyAccessTokenOutput"
                }
              }
            }
          }
        },
        "tags": [
          "OAuth Cal provider"
        ]
      }
    },
...
  },
  "info": {
    "title": "OneHash Cal v2 API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "AuthUrlData": {
        "type": "object",
        "properties": {
          "authUrl": {
            "type": "string"
          }
        },
        "required": [
          "authUrl"
        ]
      },
      "GcalAuthUrlOutput": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success",
            "enum": [
              "success",
              "error"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/AuthUrlData"
          }
        },
        "required": [
          "status",
          "data"
        ]
      },
      "GcalSaveRedirectOutput": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          }
        },
        "required": [
          "url"
        ]
      },
...
      "ProviderVerifyClientOutput": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success",
            "enum": [
              "success",
              "error"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "ProviderVerifyAccessTokenOutput": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success",
            "enum": [
              "success",
              "error"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}